Massive class-based overhaul
In this update, I have retooled much of the code to store data about cards in a Class, instead of reading it directly from permanently-stored SQL results, and everything is based off the card's ID instead of an arbitrary index in a number of parallel arrays. This does not have much of an effect on the end user, but makes the code much nicer to work with.

I also changed a few things as I was going through everything. Most notably, almost every command now supports languages - if it didn't already, it now does by typing the language after a "|".

The version number doesn't matter for much of anything because I don't actually have Bastion on npm - why would I, it's not a module - but I feel proud enough of this change to declare a totally arbitrary version 2.

I've given everything a quick test, but as with any big change there's a chance something slipped under the radar, so don't be shy about reporting any bugs you find.