Fix conflict issue and improve trivia
"Conflict issue" referring to the "Beast"-Warrior thing. I worry I've slowed down randFilterCheck implementing the fix though...

Trivia improvements are a try catch block and modification to work the same way as randcard and matches, rather than the janky way it for some reason kept.

This fork is now finally complete and I'm ready to merge it.