Fix edge case in card profiles
If a card had effect text that was exactly 1024 characters long, except then it had some trailing spaces, those spaces would be given their own section in the card profile, which would be treated as blank and cause the embed to be invalid. Occurs with Odd-Eyes Lancer Dragon (Anime) at time of writing.